--- a/.travis.yml
+++ b/.travis.yml
@@ -456,6 +456,7 @@ before_install:
   - |
     if [[ "$TRAVIS_OS_NAME" == "osx" ]]; then 
       brew update
+      brew install icu4c
     fi
   - |
     if [[ "${GOOS}" != "windows" ]]; then