--- a/.travis.yml
+++ b/.travis.yml
@@ -19,61 +19,12 @@ env:
     - GOTAGS=sqlite_vtable
 
 go:
-  - 1.7.x
   - 1.8.x
   - 1.9.x
   - 1.10.x
-  - master
 
 matrix:
-  allow_failures:
-    - go: master
-  exclude:
-    - os: osx
-      go: 1.7.x
   include:
-    - os: linux
-      go: 1.7.x
-      env: GOTAGS= GOOS=windows GOARCH=386
-    - os: linux
-      go: 1.7.x
-      env: GOTAGS= GOOS=windows GOARCH=amd64
-    - os: linux
-      go: 1.7.x
-      env: GOTAGS=libsqlite3 GOOS=windows GOARCH=386
-    - os: linux
-      go: 1.7.x
-      env: GOTAGS=libsqlite3 GOOS=windows GOARCH=amd64
-    - os: linux
-      go: 1.7.x
-      env: GOTAGS="sqlite_allow_uri_authority sqlite_app_armor sqlite_foreign_keys sqlite_fts5 sqlite_icu sqlite_introspect sqlite_json sqlite_secure_delete sqlite_see sqlite_stat4 sqlite_userauth" GOOS=windows GOARCH=386
-    - os: linux
-      go: 1.7.x
-      env: GOTAGS="sqlite_allow_uri_authority sqlite_app_armor sqlite_foreign_keys sqlite_fts5 sqlite_icu sqlite_introspect sqlite_json sqlite_secure_delete sqlite_see sqlite_stat4 sqlite_userauth" GOOS=windows GOARCH=amd64
-    - os: linux
-      go: 1.7.x
-      env: GOTAGS=sqlite_trace GOOS=windows GOARCH=386
-    - os: linux
-      go: 1.7.x
-      env: GOTAGS=sqlite_trace GOOS=windows GOARCH=amd64
-    - os: linux
-      go: 1.7.x
-      env: GOTAGS=sqlite_vacuum_full GOOS=windows GOARCH=386
-    - os: linux
-      go: 1.7.x
-      env: GOTAGS=sqlite_vacuum_full GOOS=windows GOARCH=amd64
-    - os: linux
-      go: 1.7.x
-      env: GOTAGS=sqlite_vacuum_incr GOOS=windows GOARCH=386
-    - os: linux
-      go: 1.7.x
-      env: GOTAGS=sqlite_vacuum_incr GOOS=windows GOARCH=amd64
-    - os: linux
-      go: 1.7.x
-      env: GOTAGS=sqlite_vtable GOOS=windows GOARCH=386
-    - os: linux
-      go: 1.7.x
-      env: GOTAGS=sqlite_vtable GOOS=windows GOARCH=amd64
     - os: linux
       go: 1.8.x
       env: GOTAGS= GOOS=windows GOARCH=386