--- a/sqlite3_opt_fts5.go
+++ b/sqlite3_opt_fts5.go
@@ -2,7 +2,7 @@
 //
 // Use of this source code is governed by an MIT-style
 // license that can be found in the LICENSE file.
-// +build fts5
+// +build sqlite_fts5 fts5
 
 package sqlite3