--- a/sqlite3_opt_icu.go
+++ b/sqlite3_opt_icu.go
@@ -2,7 +2,7 @@
 //
 // Use of this source code is governed by an MIT-style
 // license that can be found in the LICENSE file.
-// +build icu
+// +build sqlite_icu icu
 
 package sqlite3