--- a/sqlite3_opt_introspect.go
+++ b/sqlite3_opt_introspect.go
@@ -2,7 +2,7 @@
 //
 // Use of this source code is governed by an MIT-style
 // license that can be found in the LICENSE file.
-// +build introspect
+// +build sqlite_introspect
 
 package sqlite3