--- a/sqlite3_opt_json1.go
+++ b/sqlite3_opt_json1.go
@@ -2,7 +2,7 @@
 //
 // Use of this source code is governed by an MIT-style
 // license that can be found in the LICENSE file.
-// +build json1
+// +build sqlite_json sqlite_json1 json1
 
 package sqlite3