--- a/sqlite3_opt_stat4.go
+++ b/sqlite3_opt_stat4.go
@@ -2,7 +2,7 @@
 //
 // Use of this source code is governed by an MIT-style
 // license that can be found in the LICENSE file.
-// +build stat4
+// +build sqlite_stat4
 
 package sqlite3