--- a/sqlite3_opt_usleep.go
+++ b/sqlite3_opt_usleep.go
@@ -1,0 +1,13 @@
+// +build cgo
+// +build windows
+
+// Copyright (C) 2018 G.J.R. Timmer <gjr.timmer@gmail.com>.
+//
+// Use of this source code is governed by an MIT-style
+// license that can be found in the LICENSE file.
+
+package sqlite3
+
+/*
+
+/*