--- a/sqlite3_trace.go
+++ b/sqlite3_trace.go
@@ -2,7 +2,7 @@
 //
 // Use of this source code is governed by an MIT-style
 // license that can be found in the LICENSE file.
-// +build trace
+// +build sqlite_trace trace
 
 package sqlite3