… sqlite3_opt_fts3_test.go