[sqlite3.go](./sqlite3.go#L12) defines the following:

Line: 12
```go
#cgo CFLAGS: -DSQLITE_ENABLE_RTREE -DSQLITE_THREADSAFE=1 -DHAVE_USLEEP=1
```
This means `usleep` is activated even for Windows OS.

However usleep is a *nix command and not available on Windows.
When usleep is not available SQLite3 waits for 1 second.

This implementation adds the `usleep` command for Windows to increases performance.

[Full fix review](https://github.com/GJRTimmer/go-sqlite3/blob/7c89b38a2da041bb3bd4f5db59151ea430c5b1f0/sqlite3_usleep_windows.go)

BenchMark

| Key | Value |
|-----|--------|
|OS | Windows 10 Build: 17134.48 |
| CPU | Intel: 2.8 GHz |
| Cores | 4 |
| Memory | 16 GB |

BenchMark before fix.
```bash
BenchmarkExec           100000      40339 req/s
BenchmarkQuery           10000       3932 req/s
BenchmarkParams           5000       3367 req/s
BenchmarkStmt             5000       4010 req/s
BenchmarkRows                1          1 req/s
BenchmarkStmtRows         1000        447 req/s
```

BenchMark after fix.
```bash
BenchmarkExec           100000      77720 req/s
BenchmarkQuery           10000       6751 req/s
BenchmarkParams          10000       6579 req/s
BenchmarkStmt            10000       6878 req/s
BenchmarkRows                1          1 req/s
BenchmarkStmtRows         1000        632 req/s
```

This PR is based upon #563 please merge #563 before merging this.

@mattn Please remove you stale branch `add-usleep`.