@GJRTimmer 2018-05-25
Build successful; coverage failed probably because it was a decrease apparently because of the added file.

Can be merged after #563